+++ kolibri-fleet-zio/src/main/scala/de/awagen/kolibri/fleet/zio/taskqueue/negotiation/persistence/writer/FileStorageNodeStateWriter.scala
@@ -22,7 +22,7 @@ import de.awagen.kolibri.fleet.zio.metrics.Metrics
 import de.awagen.kolibri.fleet.zio.taskqueue.negotiation.format.FileFormats.NodeHealthFileNameFormat
 import de.awagen.kolibri.fleet.zio.taskqueue.negotiation.state.NodeUtilizationStates.NodeUtilizationStatesImplicits.nodeUtilizationStateFormat
 import de.awagen.kolibri.fleet.zio.taskqueue.negotiation.state.NodeUtilizationStates._
-import de.awagen.kolibri.fleet.zio.taskqueue.negotiation.state.ProcessingStateUtils
+import de.awagen.kolibri.fleet.zio.taskqueue.negotiation.utils.DateUtils
 import de.awagen.kolibri.storage.io.writer.Writers.Writer
 import spray.json.enrichAny
 import zio.{Task, ZIO}
@@ -41,7 +41,7 @@ case class FileStorageNodeStateWriter(writer: Writer[String, String, _]) extends
       nonHeapMax <- Metrics.CalculationsWithMetrics.getMaxNonHeapMemory
       numProcessors <- Metrics.CalculationsWithMetrics.getAvailableProcessors
     } yield NodeUtilizationState(
-      ProcessingStateUtils.timeInMillisToFormattedTime(System.currentTimeMillis()),
+      DateUtils.timeInMillisToFormattedTime(System.currentTimeMillis()),
       AppProperties.config.node_hash,
       CpuInfo(numProcessors, avgLoad),
       MemoryInfo(heapUsed, heapCommitted, heapMax),
